Checking

+++ Lesson4/4.ACT1.py
@@ -1,10 +1,10 @@
 # Asking the user to give heights
 print("Please eter the heights of all the trees in inches below.")
-height1=int(input(" Please enter the height of tree 1: "))
-height2=int(input(" Please enter the height of tree 2: "))
-height3=int(input(" Please enter the height of tree 3: ")) 
-height4=int(input(" Please enter the height of tree 4: "))
-height5=int(input(" Please enter the height of tree 5: "))
+height1=float(input(" Please enter the height of tree 1: "))
+height2=float(input(" Please enter the height of tree 2: "))
+height3=float(input(" Please enter the height of tree 3: ")) 
+height4=float(input(" Please enter the height of tree 4: "))
+height5=float(input(" Please enter the height of tree 5: "))
 
 # Summing the heights of the trees
 sum = height1 + height2 + height3 + height4 + height5
